ISSUE-#1: feature/add_collector 例外型の定義をアクセサモジュールに移動
全てのリポジトリの例外処理に利用するので抽象モジュールに移動

アクセサモジュールを全て公開する例外クラスを全て利用可能にする
