+++ trend-collector/trendcollector/libs/services/__init__.py
@@ -1,3 +1,3 @@
-from .accessor import TrendAccessor, TwitterAccountAccessor
+from .accessor import *
 from .collector import TwitterCollector
 from .custom_exception import CustomException
